+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 * 1.0.0
   - Initial release
+
+* 1.0.1
+  - Improve plugin to support multiple persistence data
+  - Change config strucute to be the same for logstash-filter-jdbc